Fix compilation issues in LMS implementation
- Fix nested namespace from #include inside NAMESPACE_BEGIN
- Fix SecureWipeBuffer calls to use .data()/.size() on SecByteBlock
- Fix CRYPTOPP_CONSTANT macro with template comma in SIGNATURE_LENGTH
- Move MakeOTSParams/MakeLMSParams helpers to lms.cpp (incomplete type)
All tests pass: store contract, keygen, sign/verify, exhaustion.
